+++ src/pages/Home/components/HomeHeader/index.tsx
@@ -3,7 +3,7 @@ import styles from './index.module.css'
 
 import Skeleton, { SkeletonTheme } from 'react-loading-skeleton';
 import 'react-loading-skeleton/dist/skeleton.css';
-import { CircularProgress } from '@mui/material';
+// import { CircularProgress } from '@mui/material';
 
 const HomeHeader: React.FC = () => {
   const homeHeaderRef = useRef<HTMLVideoElement | null>(null);
@@ -60,7 +60,8 @@ const HomeHeader: React.FC = () => {
         {isLoading ? (
           <>
             <div className={styles.headerLoading}>
-              <CircularProgress size={64} sx={{ color: 'var(--color-text-primary)' }} />
+              {/* <CircularProgress size={64} sx={{ color: 'var(--color-text-primary)' }} /> */}
+              <img className="spinnerGif" src="/images/symb/loading.gif" alt="" width={64} height={64} />
             </div>
           </>
         ) : (