+++ src/pages/Prod/components/ProdHeader/index.tsx
@@ -3,7 +3,7 @@ import styles from './index.module.css';
 
 import Skeleton, { SkeletonTheme } from 'react-loading-skeleton';
 import 'react-loading-skeleton/dist/skeleton.css';
-import { CircularProgress } from '@mui/material';
+// import { CircularProgress } from '@mui/material';
 
 const ProdHeader: React.FC = () => {
   const prodHeaderRef = useRef<HTMLVideoElement | null>(null);
@@ -58,7 +58,8 @@ const ProdHeader: React.FC = () => {
         {isLoading ? (
           <>
             <div className={styles.headerLoading}>
-              <CircularProgress size={64} sx={{ color: 'var(--color-text-primary)' }} />
+              {/* <CircularProgress size={64} sx={{ color: 'var(--color-text-primary)' }} /> */}
+              <img className="spinnerGif" src="/images/symb/loading.gif" alt="" width={64} height={64} />
             </div>
           </>
         ) : (